HTML Electricity Input Complete
User Story MSD424OL2-6
https://jira.infotech.scu.edu.au/browse/MSD424OL2-6

Note: I was having issues with using Notepad ++ for my previous 2 x commits where it would indent to far the <li> tag on line 15 and 21.  When editing the file it would not show it this way.  For this commit i used Microsoft Visual Studio and i was able to fix up the previous issues with the indent.